user can edit and delete own order. if user had mistake with ordering taxi he can edit order or delete it.

+++ CabManagementSystem/Controllers/HomeController.cs
@@ -18,6 +18,20 @@ namespace CabManagementSystem.Controllers
         {
             user.ID = HttpContext.Session.GetString("userID") is not null
                 ? new(HttpContext.Session.GetString("userID")) : new();
+            user.Order.UserID = user.ID;
+            bool predicateForExistingRow = orderContext.Orders.Any(x => x.UserID == user.ID);
+            user.Order = new()
+            {
+                UserID = user.ID,
+                PhoneNumber = predicateForExistingRow ? orderContext.Orders.First(x => x.UserID == user.ID).PhoneNumber : string.Empty,
+                Description = predicateForExistingRow ? orderContext.Orders.First(x => x.UserID == user.ID).Description : string.Empty,
+                Address = predicateForExistingRow ? orderContext.Orders.First(x => x.UserID == user.ID).Address : string.Empty,
+                DriverName = predicateForExistingRow ? orderContext.Orders.First(x => x.UserID == user.ID).DriverName : string.Empty,
+                ID = predicateForExistingRow ? orderContext.Orders.First(x => x.UserID == user.ID).ID : new()
+
+            };
+            HttpContext.Session.SetString("orderID", user.Order.ID.ToString());
+            HttpContext.Session.SetString("DriverName", user.Order.DriverName);
 
             return View(user);
         }
@@ -33,25 +47,39 @@ namespace CabManagementSystem.Controllers
             user.Order.UserID = HttpContext.Session.GetString("userID") is not null
                 ? new(HttpContext.Session.GetString("userID")) : new();
 
+            if (!applicationContext.IsAuthanticated(user.ID) && orderContext.AlreadyOrder(user.ID))
+                return RedirectToAction("Index", "Home");
+
             orderContext.CreateOrder(user.Order);
-            return RedirectToAction("Index", "Home");
-        }
-
-        [HttpPost, Route("DeleteOrder")]
-        public async Task<IActionResult> DeleteOrder(UserModel user)
-        {
-            user.Order.UserID = HttpContext.Session.GetString("userID") is not null
-                ? new(HttpContext.Session.GetString("userID")) : new();
-
-            orderContext.DeleteOrder(user.Order);
             return RedirectToAction("Index", "Home");
         }
 
         [HttpPost, Route("EditOrder")]
         public async Task<IActionResult> EditOrder(UserModel user)
         {
+            if (!applicationContext.IsAuthanticated(user.ID) && orderContext.AlreadyOrder(user.ID))
+                return RedirectToAction("Index", "Home");
+
+            user.Order.UserID = HttpContext.Session.GetString("userID") is not null
+                    ? new(HttpContext.Session.GetString("userID")) : new();
+            user.Order.ID = HttpContext.Session.GetString("orderID") is not null
+                    ? new(HttpContext.Session.GetString("orderID")) : new();
+            orderContext.UpdateOrder(user.Order);
+            return RedirectToAction("Index", "Home");
+        }
+
+        [HttpPost, Route("DeleteOrder")]
+        public async Task<IActionResult> DeleteOrder(UserModel user)
+        {
+            if (!applicationContext.IsAuthanticated(user.ID) && orderContext.AlreadyOrder(user.ID))
+                return RedirectToAction("Index", "Home");
+
             user.Order.UserID = HttpContext.Session.GetString("userID") is not null
                 ? new(HttpContext.Session.GetString("userID")) : new();
+            user.Order.ID = HttpContext.Session.GetString("orderID") is not null
+                ? new(HttpContext.Session.GetString("orderID")) : new();
+            user.Order.DriverName = HttpContext.Session.GetString("DriverName") is not null
+                ? new(HttpContext.Session.GetString("DriverName")) : string.Empty;
 
             orderContext.DeleteOrder(user.Order);
             return RedirectToAction("Index", "Home");
